🚧 J459C2 task: apply external agent result
