+++ .agentplane/tasks/202608200903-J459C2/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202608200903-J459C2/make-task-execution-authority-local-and-direct-e",
   "created_at": "2026-08-20T15:49:58.682Z",
-  "diffstat_sha256": "sha256:841dfa9c1694d30cfe61135f45023f33cb0af004ddafef180c02ddfac382a064",
+  "diffstat_sha256": "sha256:458af4749e55321563f889cdf3332175d34bb0e9ad11ac7ee58a771e1f36d59b",
   "last_verified_at": "2026-08-20T19:07:59.191Z",
   "last_verified_diffstat_sha256": "sha256:841dfa9c1694d30cfe61135f45023f33cb0af004ddafef180c02ddfac382a064",
   "pr_number": 4851,
