+++ packages/agentplane/src/commands/task/finish.close-tail.unit.test.ts
@@ -11,7 +11,7 @@ const mocks = vi.hoisted(() => ({
   cmdCommit: vi.fn(),
   ensureReconciledBeforeMutation: vi.fn(),
   loadCommandContext: vi.fn(),
-  loadTaskFromContext: vi.fn(),
+  loadTaskFromContext: vi.fn<(opts: { taskId: string }) => Promise<TaskData>>(),
   backendIsLocalFileBackend: vi.fn(),
   getTaskStore: vi.fn(),
   readCommitInfo: vi.fn(),
@@ -23,17 +23,7 @@ const mocks = vi.hoisted(() => ({
   tryLookupExistingGithubPrByBranch: vi.fn(),
   tryLookupExistingGithubPrByBranchPrefix: vi.fn(),
 }));
-vi.mock("./finish-closeout-journal.js", () => ({
-  openFinishCloseoutJournal: vi.fn().mockResolvedValue({
-    path: "/tmp/finish-closeout.json",
-    journal: { state: "prepared" },
-  }),
-  advanceFinishCloseoutJournal: vi.fn(
-    ({ journal, state }: { journal: Record<string, unknown>; state: string }) =>
-      Promise.resolve({ ...journal, state }),
-  ),
-  markFinishCloseoutRecoveryRequired: vi.fn().mockResolvedValue(undefined),
-}));
+vi.mock("./finish-closeout-journal.js", () => import("./finish-closeout-journal.testkit.js"));
 vi.mock("../guard/impl/comment-commit.js", () => ({
   commitFromComment: mocks.commitFromComment,
 }));
@@ -218,9 +208,7 @@ describe("task finish close-tail", () => {
 
   beforeEach(() => {
     for (const mock of Object.values(mocks)) mock.mockReset();
-    mocks.loadTaskFromContext.mockImplementation(({ taskId }: { taskId: string }) =>
-      Promise.resolve(mkTask({ id: taskId, status: "DOING", tags: ["spike"] })),
-    );
+    mocks.loadTaskFromContext.mockResolvedValue(mkTask({ status: "DOING", tags: ["spike"] }));
 
     mocks.backendIsLocalFileBackend.mockReturnValue(false);
     mocks.readCommitInfo.mockResolvedValue({ hash: "hc", message: "mc" });
