+++ packages/agentplane/src/commands/task/verify-record.unit.test.ts
@@ -1,15 +1,15 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import type { TaskBackend, TaskData } from "../../backends/task-backend.js";
-import type { TaskExecutionContext } from "../../runtime/task-execution-context/index.js";
 import { exitCodeForError } from "../../cli/exit-codes.js";
 import { CliError } from "../../shared/errors.js";
-import {
-  makeTaskBackendDouble,
-  makeTaskCommandContext,
-  makeTaskFixture,
-} from "@agentplane/testkit/task";
 import type { CommandContext } from "../shared/task-backend.js";
+import {
+  makeVerificationExecutionContext as executionContext,
+  makeVerifyRecordContext as mkCtx,
+  makeVerifyRecordTask as mkTask,
+  makeWriteThroughVerificationBackend as makeWriteThroughBackend,
+} from "./verify-record.testkit.js";
 
 const mocks = vi.hoisted(() => ({
   readFile: vi.fn<(p: string, enc: string) => Promise<string>>(),
@@ -89,53 +89,6 @@ vi.mock("../shared/task-store.js", async (importOriginal) => {
     }),
   };
 });
-
-function mkTask(overrides: Partial<TaskData>): TaskData {
-  return makeTaskFixture({ status: "DONE", owner: "me", ...overrides });
-}
-
-function mkCtx(overrides?: Partial<CommandContext>): CommandContext {
-  return makeTaskCommandContext({
-    taskBackend: makeTaskBackendDouble(),
-    overrides,
-    configureConfig: (config) => {
-      // Ensure the doc always includes Verification + Verify Steps for extraction/updates.
-      config.tasks.doc.required_sections = ["Summary", "Verify Steps", "Verification"];
-    },
-  });
-}
-
-function executionContext(ctx: CommandContext, taskId = "T-1"): TaskExecutionContext {
-  const mode = ctx.config.workflow_mode === "branch_pr" ? "branch_pr" : "direct";
-  return {
-    schema_version: 1,
-    primary_task_id: taskId,
-    task_ids: [taskId],
-    repository_mode: mode,
-    selected_mode: mode,
-    requested_mode: mode,
-    route_source: "execution_contract",
-    reason_codes: [],
-    base_ref: "main",
-    base_sha: "a".repeat(40),
-    authoritative_task_source: "base_checkout",
-  };
-}
-
-function makeWriteThroughBackend(opts: {
-  getTaskDoc: () => Promise<string>;
-  writeTask: TaskBackend["writeTask"];
-}): TaskBackend {
-  let persisted: TaskData | null = null;
-  return makeTaskBackendDouble({
-    getTask: () => Promise.resolve(persisted && structuredClone(persisted)),
-    getTaskDoc: opts.getTaskDoc,
-    writeTask: async (task, writeOptions) => {
-      await opts.writeTask(task, writeOptions);
-      persisted = structuredClone(task);
-    },
-  });
-}
 
 describe("task verify record (unit)", () => {
   beforeEach(() => {
